Add generate-release-notes flag to release subcommand
The GitHub REST API for releases has the option to generate the name and
description from the merged PRs since the last release. This commit also
adds the option to the release subcommand.

This should not be a backwards breaking change, because the name and the
description still defaults to the tag, if the -g argument is not given.
If the -g argument is given, the name will be empty and GitHub will
display the tag in the web interface and the description will be the
automatically created description by GitHub.